+++ src/layout/MainLayout/styles.module.css
@@ -1,5 +1,12 @@
 .main_layout{
-        display: grid;
+         display: grid;
     grid-template-columns: auto 1fr;
+     /* height: 100vh;  */
+
+
+}
+.main_content{
     height: 100vh;
+    width: 100%;
+    background-color: #eee;
 }
